requirements.txt
psutil
pystray
Pillow
pywin32